refactor(graphql): rename contact-create.ts to contact.ts to match object name
