+++ core/api/src/graphql/public/types/payload/contact.ts
@@ -1,10 +1,10 @@
-import Contact from "../object/contact-create"
+import Contact from "../object/contact"
 
 import IError from "@/graphql/shared/types/abstract/error"
 import { GT } from "@/graphql/index"
 
 const ContactUpdateOrCreatePayload = GT.Object({
-  name: "ContactUpdateOrCreatePayload",
+  name: "ContactPayload",
   fields: () => ({
     errors: {
       type: GT.NonNullList(IError),
